Fix DECIMAL type incompatibility by changing to DOUBLE in test schema
Fixed the test schema error where gocql could not marshal float64 (19.99)
into a DECIMAL column type.

Changes:
1. internal/action/main_test.go:
   - Changed products table price column from DECIMAL to DOUBLE
   - DOUBLE is directly compatible with Go's float64 type
   - No changes needed to the insert statement

2. internal/action/README_TESTS.md:
   - Updated documentation to reflect DOUBLE type for price column

Technical Details:
- gocql does not automatically marshal float64 to DECIMAL type
- DECIMAL requires special handling with inf.Dec type
- DOUBLE is the standard CQL type for floating-point numbers
- DOUBLE maps directly to Go's float64, making it easier to use in tests

Error Fixed:
"can not marshal float64 into decimal"

The test suite now successfully creates the schema and inserts test data.

+++ internal/action/main_test.go
@@ -239,7 +239,7 @@ func setupTestSchema(hostPort string) error {
 		CREATE TABLE IF NOT EXISTS products (
 			id UUID PRIMARY KEY,
 			name TEXT,
-			price DECIMAL,
+			price DOUBLE,
 			stock INT
 		)
 	`).Exec(); err != nil {
